+++ src/main/ts/view/value/graph.ts
