+++ src/main/ts/view/value/monitor.ts
