+++ src/main/ts/view/value/multi-log.ts
