+++ src/main/ts/view/value/single-log.ts
